requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124

# Validated working stack for this repo: Python 3.12.3 + CUDA 12.4.
# The torchao pin is intentional; newer torchao builds emit an incompatible
# torch warning with the validated torch 2.6.0+cu124 / transformers 5.3.0 stack.

# Core Python / PyTorch
torch==2.6.0+cu124
torchvision==0.21.0+cu124
torchaudio==2.6.0+cu124

# Hugging Face ecosystem
transformers==5.3.0
datasets==4.3.0
peft==0.18.1
trl==0.24.0
accelerate==1.13.0
huggingface_hub==1.8.0

# Quantization / kernels
bitsandbytes==0.49.2
unsloth==2026.3.17
torchao==0.12.0

# Experiment tracking
wandb==0.19.1
tensorboard==2.20.0

# Evaluation
evaluate==0.4.3
rouge-score==0.1.2
openai==2.30.0

# Export and utilities
gguf==0.18.0
py7zr==1.1.0
pyzstd==0.19.1
python-dotenv==1.0.1
pyyaml==6.0.3
requests==2.32.3
beautifulsoup4==4.12.3
pypdf==6.9.2
numpy==2.3.5
sentencepiece==0.2.1
protobuf==5.29.6
